% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineages.R
\name{get_lineages}
\alias{get_lineages}
\title{Get full lineage information for taxa.}
\usage{
get_lineages(
  taxids,
  ranks = c("Domain", "Kingdom", "Phylum", "Class", "Order", "Superfamily", "Family",
    "Subfamily", "Genus"),
  tdb = NULL
)
}
\arguments{
\item{taxids}{A numeric vector of taxids to pull taxonomic information for.}

\item{ranks}{A character vector of taxonomic ranks to return.}

\item{tdb}{Optional. A taxonomy table in `insect` package format}
}
\value{
A dataframe with one row for every taxid and one column for each rank.
}
\description{
Get full lineage information for taxa.
}
\details{
This function takes a vector of taxids and uses the insect package to pull the full lineage information for each taxid.
}
\author{
Shaun Wilkinson and Susan Welsh
}
