% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_wilderdata}
\alias{get_wilderdata}
\title{Access eDNA data from the Wilderlab API.}
\usage{
get_wilderdata(tb, key = NULL, secret = NULL, xapikey = NULL, JobID = NULL)
}
\arguments{
\item{tb}{a character string specifying the table required. Accepted values are 'jobs', 'samples', 'taxa' and 'records'}

\item{key}{a string specifying the API access key for the client account. Not required for taxa table.
Please contact info at wilderlab.co if you would like access keys generated for your account.}

\item{secret}{a string specifying the API secret access key for the client account. Not required for taxa table.
Please contact info at wilderlab.co if you would like access keys generated for your account.}

\item{JobID}{a 6 digit integer specifying a Wilderlab job number. Required for accessing the records table.}

\item{xapi}{a string specifying the X-API-Key value for the client account. Not required for taxa table.
Please contact info at wilderlab.co if you would like access keys generated for your account.}
}
\description{
Wrapper functions for creating URLs and authorizaton headers to download job, sample, taxa and record information
from the connect.wilderlab API.
}
\details{
The Wilderlab API is designed for clients to access up-to-date eDNA data for their internal data storage
  platforms and geospatial applications. Clients can access their job, sample, taxon and eDNA records data at any time
  by querying the API with a valid URL and authorisation header. The get_wilderdata function is a wrapper that enables
  these URLs and headers to be compiled with minimal effort.
}
\examples{
\donttest{
    key <- "AKIATVYXGCYLWADFJVEX"
    secret <- "SiDvZFUFXlCXK/jeBtHrfRPWMmb8veW6q5+ULuyx"
    xapikey <- "7CCm580l5vgeKbalwIEy565uFhbEudTauAq80B38"
    jobs <- get_wilderdata("jobs", key = key, secret = secret, xapikey = xapikey)
    samples <- get_wilderdata("samples", key = key, secret = secret, xapikey = xapikey)
    taxa <- get_wilderdata("taxa", key = key, secret = secret, xapikey = xapikey)
    records <- vector(mode = "list", length = nrow(jobs))
    for(i in seq_along(records)){
      records[[i]] <- get_wilderdata("records", JobID = jobs$JobID[i], key = key, secret = secret, xapikey = xapikey)
    }
    records <- do.call("rbind", records)
 }
}
\seealso{
Full tutorial available at \link{wilderlab.co/api-instructions}
}
\author{
Shaun Wilkinson and Susan Welsh
}
