% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eDNA.R
\name{read_eDNA}
\alias{read_eDNA}
\title{Read in Wilderlab results spreadsheets.}
\usage{
read_eDNA(filepath = file.choose())
}
\arguments{
\item{filepath}{A character string specifying path to the Wilderlab results file. Only .xlsx files are allowed. If
blank, an interactive dialog box will be presented to the user.}
}
\value{
A named list of tibbles of each sheet in the xlsx file. Note the metadata sheet is separated into job and samples metadata.
}
\description{
Wrapper functions for creating URLs and authorizaton headers to download job, sample, taxa and record information
from the connect.wilderlab API.
}
\details{
This function is designed for easily importing Wilderlab results spreadsheets into an R environment.
}
\author{
Shaun Wilkinson and Susan Welsh
}
